python/tkinter_pdf_text_extractor/requirements.txt
colour==0.1.5
Pillow==8.3.2
PyPDF2==1.26.0
tk==0.1.0
tkmacosx==1.0.3
